TST, MAINT: Re-ordered testing config/fixtures - Separated sequential / distributed tests related fixtures
