+++ tests/component_integration/solvers/dense_solvers/sequential/test_logdet.py
@@ -1,6 +1,9 @@
 # Copyright 2024-2025 DALIA authors. All rights reserved.
 
+import pytest
 
+
+@pytest.mark.mpi_skip()
 def test_logdet_correctness(
     reference_logdet,
     allclose_floats,
