+++ tests/component_integration/solvers/structured_solvers/sequential/test_factorize.py
@@ -1,6 +1,9 @@
 # Copyright 2024-2025 DALIA authors. All rights reserved.
 
+import pytest
 
+
+@pytest.mark.mpi_skip()
 def test_factorize_correctness(
     allclose_dense_structured,
     reference_cholesky,
